Move h.admin.worker to h.tasks.admin
