+++ h/admin/views/users.py
@@ -9,8 +9,8 @@ from pyramid.view import view_config
 
 from h import models
 from h.accounts.events import ActivationEvent
-from h.admin import worker
 from h.admin.services.user import UserRenameError
+from h.tasks.admin import rename_user
 from memex import storage
 from h.i18n import TranslationString as _
 
@@ -87,7 +87,7 @@ def users_rename(request):
         svc = request.find_service(name='rename_user')
         svc.check(new_username, request.auth_domain)
 
-        worker.rename_user.delay(user.id, new_username)
+        rename_user.delay(user.id, new_username)
 
         request.session.flash(
             'The user "%s" will be renamed to "%s" in the backgroud. Refresh this page to see if it\'s already done' %
